+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Made APC store setMulti with integer keys
 - Short-circuit getting/setting/deleting multiple keys
 - Make sure SQLite treats numerical keys as strings, not numbers
+- Support numerical keys with Couchbase
 
 
 ## [1.4.2] - 2017-02-07